+++ mjml/core/registry.py
@@ -9,7 +9,7 @@ components = {}
 def register_core_components():
     from ..elements import (MjAccordion, MjAccordionElement, MjAccordionText, MjAccordionTitle,
         MjButton, MjText, MjSection, MjColumn, MjBody,
-        MjGroup, MjHero, MjImage, MjNavbar, MjNavbarLink, MjDivider, MjSpacer, MjTable, MjRaw, MjWrapper)
+        MjGroup, MjHero, MjImage, MjNavbar, MjNavbarLink, MjDivider, MjSocial, MjSocialElement, MjSpacer, MjTable, MjRaw, MjWrapper)
     from ..elements.head import (MjAttributes, MjBreakpoint, MjFont, MjHead, MjPreview, MjStyle,
         MjTitle)
 
@@ -31,6 +31,8 @@ def register_core_components():
         MjRaw,
         MjNavbar,
         MjNavbarLink,
+        MjSocial,
+        MjSocialElement,
         MjSpacer,
         MjWrapper,
         # --- head components ---